requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiostream==0.5.2
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0
fastapi==0.115.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
grandalf==0.8
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langgraph==0.2.39
langgraph-checkpoint==2.0.1
langgraph-sdk==0.1.33
langsmith==0.1.136
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
modal==0.63.57
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.10.0
numpy==1.26.4
openai==1.52.1
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
propcache==0.2.0
protobuf==4.25.5
pyarrow==17.0.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydash==8.0.3
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
pytest==8.3.3
pytest-requests==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
shellingham==1.5.4
sigtools==4.0.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
streamlit==1.37.0
synchronicity==0.6.7
tenacity==8.5.0
tiktoken==0.8.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.5
trafaret==2.1.1
typer==0.12.5
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchfiles==0.24.0
yarl==1.16.0
diff-match-patch